azure_bot_sdk_demo/assistant_bot/requirements.txt
botbuilder-core
asyncio
aiohttp
cookiecutter==1.7.0
openai